pyproject.toml
[project]
name = "input-tracing"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "click>=8.1.7",
    "openai>=1.54.0",
    "pangea-sdk>=5.1.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.13.0",
    "ruff>=0.7.2",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # ruff defaults
    "E4",
    "E7",
    "E9",
    "F",

    "I",   # isort
    "PTH", # flake8-use-pathlib
]
